requirements.txt
django==2.2.0
psycopg2
djangorestframework==3.7.1